% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confirmed.fromMatrix.R
\name{Confirmed.fromMatrix}
\alias{Confirmed.fromMatrix}
\title{Confirmed.fromMatrix}
\usage{
Confirmed.fromMatrix(InterMatrix, GroundT)
}
\arguments{
\item{InterMatrix}{A \code{matrix} containing miRNA-mRNA interactions.
The first columns corresponds to miRNAs. The second column corresponds to mRNA.}

\item{GroundT}{A \code{list} where each element has the following structure:
\itemize{
  \item{\code{name:}}{ mRNA name.}
  \item{\code{data:}}{ Names of the miRNAs whose interactions with the mRNA have
  been experimentally confirmed.}
}}
}
\value{
A \code{matrix} containing all interactions in \code{InterMatrix} that
 are in the \code{GroundT}
}
\description{
Given a matrix with miRNA-mRNA interactions, returns a \code{matrix} with all
of those interactions that are present in the \code{GroundT}.
}
\examples{
\dontrun{
data(TCGA_BRCAdata)
data(GroundT)
test1<-PTC(miRNAs=TCGA_BRCAdata$miRs,mRNAs=TCGA_BRCAdata$mRNAs, VIM=TCGA_BRCAdata$mRNAs[,"VIM"])
aux<-InterList.toMatrix(test1$Names)
t1.Confirmed<-Confirmed.fromMatrix(aux,GroundT)
}
}
\seealso{
\link{PTC}, \link{GroundT} \link{Confirmed.fromList}
 \link{InterList.toMatrix}
}
