% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.Confirmed.R
\name{Find.Confirmed}
\alias{Find.Confirmed}
\title{Find.Confirmed}
\usage{
Find.Confirmed(miRNAsList, GroundT)
}
\arguments{
\item{miRNAsList}{A \code{list} where each element has the following structure:
\itemize{
  \item{\code{name:}}{ mRNA name.}
  \item{\code{data:}}{ Names of the miRNAs inferred as parents by PTC.}
}}

\item{GroundT}{A \code{list} where each element has the following structure:
\itemize{
  \item{\code{name:}}{ mRNA name.}
  \item{\code{data:}}{ Names of the miRNAs whose interactions with the mRNA have
  been experimentally confirmed.}
}}
}
\value{
A \code{list} where each element has the following structure:
   \item{name}{ Name of a mRNA inferred by PTC with at least one confirmed interaction.}
   \item{data}{ Names of the miRNAs whose miRNA-mRNA interactions are in the \code{GroundT}.}
}
\description{
Returns a \code{list} with all experimentally confirmed interactions
inferred by PTC.
}
\examples{
\dontrun{
data(TCGA_BRCAdata)
data(GroundT)
test1<-PTC(miRNAs=TCGA_BRCAdata$miRs,mRNAs=TCGA_BRCAdata$mRNAs, VIM=TCGA_BRCAdata$mRNAs[,"VIM"])
t1.Confirmed<-Find.Confirmed(test1$Plist$Names,GroundT)
}
}
\seealso{
\link{PTC}, \link{GroundT}
}
