% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatabyMAD.R
\name{getDatabyMAD}
\alias{getDatabyMAD}
\title{getDatabyMAD}
\usage{
getDatabyMAD(seqData, nmiR, nmR)
}
\arguments{
\item{seqData}{A list with two elements:
   \itemize{
      \item A matrix with miRNAs gene expression.
      \item A matrix with mRNAs  gene expression.
   }
Columns represent miRNAs/mRNAs and rows represent samples (time points). Columns names must
correspond to miRNAs/mRNAs names.}

\item{nmiR}{Number of miRNAs to be selected as predictor candidates.}

\item{nmR}{Number of mRNAs to be selected as target variables.}
}
\value{
A \code{list} containing samples of miRs and mRs. The elements of the \code{list} are:
   \item{\code{d}}{ The data with rows being samples and columns being miRs and mRs.}
   \item{\code{miRs}}{ The names of the selected miRs.}
   \item{\code{mRs}}{ The names of the selected mRs.}
}
\description{
Given a gene expression data including matched miRNAs-mRNAs,
   finds the \code{nmiR} miRNAs and the \code{nmR} mRNAs with the largest
   Median Absolute Deviation (MAD).
}
\examples{
data(TCGA_BRCAdata)
nmiR <- 30
nmR <- 1500
l <- getDatabyMAD(TCGA_BRCAdata, nmiR, nmR)
}
\references{
Taosheng Xu, Thuc Duy Le, Lin Liu, Ning Su,
 Rujing Wang, Bingyu Sun, Antonio Colaprico,
 Gianluca Bontempi, Jiuyong Li.
 CancerSubtypes: an R/Bioconductor package for molecular
 cancer subtype identification, validation, and visualization. Bioinformatics 33(19): 3131–3133 (2017).
 \url{https://doi.org/10.1093/bioinformatics/btx378}
}
\seealso{
\link[CancerSubtypes]{FSbyMAD}
}
\author{
Vu VH Pham <vu_viet_hoang.pham@mymail.unisa.edu.au>,
        Taosheng Xu <taosheng.x@gmail.com>
}
