% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterList.toMatrix.R
\name{InterList.toMatrix}
\alias{InterList.toMatrix}
\title{InterList.toMatrix}
\usage{
InterList.toMatrix(miRNAsList)
}
\arguments{
\item{miRNAsList}{A \code{list} where each element has the following structure:
\itemize{
  \item{\code{name:}}{ mRNA name.}
  \item{\code{data:}}{ Names of the miRNAs inferred as parents by PTC.}
}}
}
\value{
A matrix with 2 columns containing the miRNA-mRNA interactions.
}
\description{
Given a \code{list} of mRNAs containing the set of their
inferred parents (the output of \link{Extract.Parents}), it returns
a \code{matrix} with the miRNA-mRNA interactions. Each row represents an interaction.
The first column contains miRNAs. The second column contains mRNAs
}
\examples{
\dontrun{
data(TCGA_BRCAdata)
data(TScan)
seqData<-PTC.ptime(TCGA_BRCAdata,TCGA_BRCAdata$mRNAs[,"VIM"])
SelData<-PTC.GeneSel(seqData, topk_miR = 30, topk_mR = 1500)
PParents<-PTC.findPP(TScan, miRs=SelData$miRs, mRs="SORCS1")
temp <-SelData$PParents[["SORCS1"]]
SORCS1.X.TScan=SelData$d[,temp]
SORCS1.Parents<-PTC.TestInvariance(Y=SelData$d[,"SORCS1"], X=SORCS1.X.TScan)
SORCS1.Parents<-Extract.Parents(SORCS1.Parents,SelData$PParents[["SORCS1"]])
SORCS1.relMatrix<-na.omit(InterList.toMatrix(SORCS1.Parents$Plist$Names))
}
}
\seealso{

}
