% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterMatrix.toList.R
\name{InterMatrix.toList}
\alias{InterMatrix.toList}
\title{InterMatrix.toList}
\usage{
InterMatrix.toList(InterMatrix)
}
\arguments{
\item{InterMatrix}{a two column \code{matrix}. The first column corresponds to miRNA.
the second column corresponds to mRNA.}
}
\value{
A \code{list} where each element has the following structure:
   \item{\code{name:}}{ mRNA name.}
   \item{\code{data:}}{ Names of the miRNAs linked to the current mRNA}
}
\description{
Given a matrix containing miRNA-mRNA pairs,
creates a /code{list} where each element corresponds to a mRNA and contains
all miRNA linked to that mRNA.
}
\seealso{

}
