% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTC.ConfirmedInTopK.R
\name{PTC.ConfirmedInTopK}
\alias{PTC.ConfirmedInTopK}
\title{This fuction finds the number of confirmed interactions, given
a matrix of interacions miR-mR, and a Ground Truth}
\usage{
PTC.ConfirmedInTopK(
  GT,
  relationshipsM,
  topK,
  GT_miR_EMT = NULL,
  GT_mR_EMT = NULL
)
}
\arguments{
\item{GT:}{a list where each element has
name=gene, data= confirmed miRs able to bind that gene}

\item{relationshipsM:}{A matrix [n,2]}

\item{topK:}{a vector including all topK}
}
\value{
A matrix [n,3], where the additional column has a score
calculated as the mean of all context score of each miR-mR pair
}
\description{
This fuction finds the number of confirmed interactions, given
a matrix of interacions miR-mR, and a Ground Truth
}
