% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTC.findPP.R
\name{PTC.findPP}
\alias{PTC.findPP}
\title{PTC.findPP}
\usage{
PTC.findPP(TScan, miRs, mRs)
}
\arguments{
\item{TScan}{A \code{matrix} containing miRNA-mRNA interactions predicted by
TargetScan 7.0. The first and second columns contain miRNAs names from
miRBase versions 18 and 17 respectively. The third column contains the names of the target mRNAs.}

\item{miRs}{A vector containing the names of miRNAs to be verified. These names
corresponds to the names of predictor candidates.}

\item{mRs}{A vector containing the names of mRNAs to be verified. These names
corresponds to the names of target genes (response variables).}
}
\value{
A list containing the set of miRNAs that can bind each target mRNA:
}
\description{
For each mRNA, creates a set of miRNAs (plausible parents) that can biologically target that mRNA.
A total of \code{nmiR} miRNAs and \code{nmR} mRNAs with the largest gene expression Median Absolute Deviation (MAD)
are selected as predictor candidates and target mRNAs respectively.
\cr The set of plausible parents of each gene is a subset of the predictor candidates and contains those miRNA that can bind
the gene as predicted by TargetScan 7.0.
miRNAs names are converted to \code{miRBase v.21} during selection process.
}
\examples{
\dontrun{
data(TCGA_BRCAdata)
data(TScan)
seqData<-PTC.ptime(TCGA_BRCAdata,TCGA_BRCAdata$mRNAs[,"VIM"])
SelData<-PTC.GeneSel(seqData, topk_miR = 30, topk_mR = 1500)
PParents<-PTC.findPP(TScan, miRs=SelData$miRs, mRs=SelData$mRs)
}
}
\seealso{
\link{PTC},\link{PTC.GeneSel}, \link{TScan}
}
