% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTC.miRv21.R
\name{PTC.miRv21}
\alias{PTC.miRv21}
\title{PTC.miRv21}
\usage{
PTC.miRv21(miRs)
}
\arguments{
\item{miRs}{miRNAs names to be changed to version \code{miRBase v.21}}
}
\value{
\code{nmiR} A vector with the miRNAs names
(version \code{miRBase v.21})
}
\description{
Takes a vector containing miRNAs names and returns
the \code{miRBase v.21} version of them.
}
\examples{
\dontrun{
data(TCGA_BRCAdata)
miRnamesv21<-PTC.miRv21(colnames(TCGA_BRCAdata$miRs))
 }
}
\seealso{
\link{PTC}, \link[miRBaseConverter]{miRNAVersionConvert},
\link{PTC.GeneSel}
}
