% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTC.RankByContext.R
\name{PTC.RankByContext}
\alias{PTC.RankByContext}
\title{PTC.RankByTScontext}
\usage{
PTC.RankByContext(TS70context, relationshipsM)
}
\arguments{
\item{TS70context:}{A \code{matrix} containing conserved site context++ scores. The headers
of the matrix from TargetScan must be conserved when the file is imported into R.}

\item{relationshipsM:}{A matrix with two columns containing miRNAs and mRNAs names.
First column represents miRNAs and second column mRNAs (e.g the output of \link{InterList.toMatrix}).}
}
\value{
A relationship matrix with four columns [Rank, miRNA, mRNA, Score].
The more negative the score, the best the rank of the corresponding miRNA-mRNA pair.
}
\description{
This fuction uses TS70 Conserved_Site_Context_Scores (from
\url{http://www.targetscan.org/cgi-bin/targetscan/data_download.cgi?db=vert_70}
to rank a matrix of inferred miRNA-mRNA relationships.
}
\examples{
\dontrun{
data(TCGA_BRCAdata)
data(TScan)
data(TS7.0_Conserved_Site_Context_Scores)
seqData<-PTC.ptime(TCGA_BRCAdata,TCGA_BRCAdata$mRNAs[,"VIM"])
SelData<-PTC.GeneSel(seqData, topk_miR = 30, topk_mR = 1500)
PParents<-PTC.findPP(TScan, miRs=SelData$miRs, mRs="SORCS1")
temp <-SelData$PParents[["SORCS1"]]
SORCS1.X.TScan=SelData$d[,temp]
SORCS1.Parents<-PTC.TestInvariance(Y=SelData$d[,"SORCS1"], X=SORCS1.X.TScan)
SORCS1.Parents<-Extract.Parents(SORCS1.Parents,SelData$PParents[["SORCS1"]])
SORCS1.relMatrix<-na.omit(InterList.toMatrix(SORCS1.Parents$Plist$Names))
SORCS1.Summary<-PTC.RankByContext(TS7.0_Conserved_Site_Context_Scores
                                  ,SORCS1.relMatrix)
}
}
\seealso{
\link[PTC]{PTC}
}
