% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTC.R
\name{PTC}
\alias{PTC}
\title{PTC}
\usage{
PTC(miRNAs, mRNAs, VIM, nmiR=30, nmR=1500 , ngrid=2, alpha=0.02, complements = TRUE
, explore.all=TRUE, silent=TRUE)\cr

For an adequate functioning, columns names in both, \code{miRNAs} and \code{mRNAs}
matrices, must be the names of the miRNAs and mRNAs respectively.
}
\arguments{
\item{miRNAs}{A \code{matrix} containing miRNA gene expression.\cr
A total of \code{nmiR} miRNAs from this matrix are selected to be used
as predictors candidates (i.e. plausible parents).
Columns represent miRNAs, rows represent samples.}

\item{mRNAs}{A \code{matrix} containing mRNA gene expression.\cr
A total of \code{nmR} mRNAs from this matrix are
selected to be used as response variables.
Columns represent mRNAs, rows represent samples.}

\item{VIM}{VIM expression to be used for calculating VIM_Time.}

\item{nmiR}{Number of miRNAs to be selected as predictor candidates.}

\item{nmR}{Number of mRNAs to be selected as target variables.}

\item{ngrid}{Number of segments of the time series data used for creating
the different enviroments required for the statistical test.
\code{ngrid=2} by default.}

\item{alpha}{Significance level for the statistical test.
\code{alpha=0.02} by default.}

\item{complements}{If \code{TRUE} (default), each environment is compared against
its complement. If \code{FALSE} all environments are compared pairwise.}

\item{explore.all}{If \code{TRUE}(default), PTC explores all combinations of predictors
and returns the union set of all combinations that does not violate the invariance
property. If \code{FALSE} PTC returns the first set that does not violate the invariance
property. Exploration is made starting from the set of all predictors and reducing
the size of the set by one predictor at a time.}

\item{silent}{If \code{TRUE} (default), PTC displays the currently evaluated set.
If \code{FALSE}, PTC only displays the number of sets to be explored in
the current iteration.}
}
\value{
A \code{list} consisting of the following elements:
  \item{\code{Index}}{A \code{list} where each element is a target mRNA.
   For each target gene with at least one parent. The index of the parents.}
  \item{\code{names}}{A \code{list} where each element is a target mRNA.
   For each target gene with at least one parent. The name of the parents.}
  \item{\code{genes}}{The names of all target genes with at least one parent.}
  \item{\code{Summary}}{A matrix representing miRNA-mRNA regulatory interactions
  inferred by PTC. The columns of the matrix are:
       \enumerate{
          \item{\code{rank:}}{ Rank of the inferred interaction}
          \item{\code{miR:}}{ Names of miRNA (Parent)}
          \item{\code{mR:}}{ Names of mRNA (Child)}
          \item{\code{Score:}}{ Score as calculated by \link{PTC.RankByContext}}
          }
   }
}
\description{
PTC estimates the causal parents of a set of \code{nmR} mRNAs,
given a set of \code{nmiR} predictors (miRNAs). This estimation assumes
a linear model \deqn{Y_t = \beta X_t + \epsilon}
With \eqn{Y_t, X_t} the sequential
data of the target and the predictors respectively, and \eqn{\epsilon} independent and identically
distributed errors for the n time points in the time series.
\eqn{Y_t, X_t} are the sequential data obtained after a pseudotime analysis.
}
\examples{
\dontrun{
   data(TCGA_BRCAdata)
   test1<-PTC(miRNAs=TCGA_BRCAdata$miRs, mRNAs=TCGA_BRCAdata$mRNAs
          , VIM=TCGA_BRCAdata$mRNAs[,"VIM"])
}

}
\references{
A Pseudo-Temporal Causality Approach to Identifying miRNA-mRNA
Interactions During Biological Processes\cr
Andres M. Cifuentes-Bernal, Vu VH Pham, Xiaomei Li,
Lin Liu, Jiuyong Li, Thuc Duy Le \cr
bioRxiv 2020.07.07.192724;
 \url{https://doi.org/10.1101/2020.07.07.192724}
}
\seealso{
\link[PTC]{PTC.ptime}, \link[PTC]{PTC.GeneSel},
\link[PTC]{PTC.TestInvariance}.
}
\author{
Andres Mauricio Cifuentes_Bernal, Vu VH Pham, Xiaomei Li, Lin Liu, JiuyongLi and Thuc Duy Le
}
