% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_miRNAsdata.R
\docType{data}
\name{SC_miRNAsdata}
\alias{SC_miRNAsdata}
\title{Original miRNAs dataset from GSE114071}
\format{
A \code{matrix} with 2822 rows and 23 columns.
samples are represented in rows. Columns contains miRNAs names,
description and samples
}
\usage{
SC_miRNAsdata
}
\description{
File containing miRNA normalized and log2 transformed
gene expresion of 2822 miRNAs from 21 samples.
}
\references{
Wang, N., Zheng, J., Chen, Z. et al.
"Single-cell microRNA-mRNA co-sequencing reveals non-genetic heterogeneity
and mechanisms of microRNA regulation."
Nat Commun 10, 95 (2019).
https://doi.org/10.1038/s41467-018-07981-6
}
\keyword{Cell}
\keyword{Single}
\keyword{dataset,}
\keyword{miRNAs,}
