% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_mRNAsdata.R
\docType{data}
\name{SC_mRNAsdata}
\alias{SC_mRNAsdata}
\title{Original mRNAs dataset from GSE114071}
\format{
A \code{matrix} with 23285 rows and 24 columns.
samples are represented in rows (from row 2). Columns contains miRNAs names,
description and samples
}
\usage{
SC_mRNAsdata
}
\description{
GSE114071 file containing mRNA gene expresion
of 23284 mRNAs from 22 samples.
}
\references{
Wang, N., Zheng, J., Chen, Z. et al.
"Single-cell microRNA-mRNA co-sequencing reveals non-genetic heterogeneity
and mechanisms of microRNA regulation."
Nat Commun 10, 95 (2019).
https://doi.org/10.1038/s41467-018-07981-6
}
\keyword{Single_Cell}
\keyword{dataset}
\keyword{mRNAs}
