% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_BRCAdata.R
\docType{data}
\name{TCGA_BRCAdata}
\alias{TCGA_BRCAdata}
\title{miRNA-mRNA matched bulk data from TCGA BRCA project}
\format{
A \code{list} object with 2 elements: miRs and mRNAs
}
\usage{
TCGA_BRCAdata
}
\description{
File containing BRCA gene expresion matched data
of 518 miRNAs and 17403 mRNAs from 503 samples.
}
\references{
Pham, V., Zhang, J., Liu, L. et al.
"Identifying miRNA-mRNA regulatory relationships in breast cancer with invariant causal prediction."
BMC Bioinformatics 20, 143 (2019).
https://doi.org/10.1186/s12859-019-2668-x
}
\keyword{dataset}
