% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS7.0_Conserved_Site_Context_Scores.R
\docType{data}
\name{TS7.0_Conserved_Site_Context_Scores}
\alias{TS7.0_Conserved_Site_Context_Scores}
\title{TargetScan 7.0 Conserved Site Context Scores.}
\format{
A \code{matrix} with TargetScan 7.0 Conserved Site Context Scores data.
}
\usage{
TS7.0_Conserved_Site_Context_Scores
}
\description{
File containing a \code{matrix} with the TargetScan 7.0 Context++ scores
and contributions for all conserved miRNA sites,
downloaded from \url{http://www.targetscan.org/cgi-bin/targetscan/data_download.cgi?db=vert_70}
}
\references{
Agarwal, V.; Bell, G. W.; Nam, J.-W. & Bartel, D. P.
Predicting effective microRNA target sites in mammalian mRNAs
eLife, eLife Sciences Publications, Ltd, 2015, 4
}
\keyword{dataset}
