% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScan.R
\docType{data}
\name{TScan}
\alias{TScan}
\title{TargetScan 7.0}
\format{
A \code{matrix} with 2 columns
\itemize{
  \item{miRNAs names v21}
  \item{predicted target (mRNA)}
}
}
\usage{
TScan
}
\description{
File containing a \code{matrix} with miRNAs and
predicted targets from TargetScan Release 7.0. miRNA names were transformed
to their miRBase v.21 by using \link{PTC.miRv21}
}
\references{
Agarwal, V.; Bell, G. W.; Nam, J.-W. & Bartel, D. P.
Predicting effective microRNA target sites in mammalian mRNAs
eLife, eLife Sciences Publications, Ltd, 2015, 4
}
\keyword{dataset}
