% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterList.toMatrix.R
\name{InterList.toMatrix}
\alias{InterList.toMatrix}
\title{InterList.toMatrix}
\usage{
InterList.toMatrix(InterList)
}
\arguments{
\item{InterList}{A \code{list} where each element has the following structure:
\itemize{
  \item{\code{name:}}{ mRNA name.}
  \item{\code{data:}}{ Names of the miRNAs inferred as parents}
}}
}
\value{
A matrix with 2 columns containing the miRNA-mRNA interactions.
}
\description{
Given a \code{list} where each element is a mRNA containing the set of
the miRNAs (parents) such mRNA can interact with(e.g. the output of \link{Extract.Parents}),
it returns a \code{matrix} with the miRNA-mRNA interactions. Each row represents an interaction.
The first column contains miRNAs. The second column contains mRNAs
}
\examples{
\dontrun{
data(TCGA_BRCAdata)
data(TScan)
seqData<-PTC.ptime(TCGA_BRCAdata,TCGA_BRCAdata$mRNAs[,"VIM"])
SelData<-PTC.GeneSel(seqData, nmiR = 30, nmR = 1500)
PParents<-PTC.findPP(TScan, miRs=SelData$miRs, mRs="ONECUT2")
temp <-SelData$PParents[["ONECUT2"]]
ONECUT2.X.TScan=SelData$d[,temp]
set.seed(1)
ONECUT2.Parents<-PTC.TestInvariance(Y=SelData$d[,"ONECUT2", drop = F], X=ONECUT2.X.TScan)
ONECUT2.Parents<-Extract.Parents(ONECUT2.Parents,SelData$PParents["ONECUT2"])
ONECUT2.relMatrix<-InterList.toMatrix(ONECUT2.Parents$Names)
}
}
\seealso{

}
