% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTC.RankByContext.R
\name{PTC.RankByContext}
\alias{PTC.RankByContext}
\title{PTC.RankByTScontext}
\usage{
PTC.RankByContext(TS70context, relationshipsM)
}
\arguments{
\item{TS70context:}{A \code{matrix} containing conserved site context++ scores. The headers
of the matrix from TargetScan must be conserved when the file is imported into R.}

\item{relationshipsM:}{A matrix with two columns containing miRNAs and mRNAs names.
First column represents miRNAs and second column mRNAs (e.g the output of \link{InterList.toMatrix}).}
}
\value{
A relationship matrix with four columns [Rank, miRNA, mRNA, Score].
The more negative the score, the best the rank of the corresponding miRNA-mRNA pair.
}
\description{
This fuction uses TS70 Conserved_Site_Context_Scores (from
\url{http://www.targetscan.org/cgi-bin/targetscan/data_download.cgi?db=vert_70})
to rank a matrix of inferred miRNA-mRNA relationships.
}
\examples{
\dontrun{
data(TCGA_BRCAdata)
data(TScan)
data(TS7.0_Conserved_Site_Context_Scores)

seqData<-PTC.ptime(TCGA_BRCAdata,TCGA_BRCAdata$mRNAs[,"VIM"])
SelData<-PTC.GeneSel(seqData, nmiR = 30, nmR = 1500)
PParents<-PTC.findPP(TScan, miRs=SelData$miRs, mRs="ONECUT2")
temp <-SelData$PParents[["ONECUT2"]]

ONECUT2.X.TScan=SelData$d[,temp]
set.seed(1)
ONECUT2.Parents<-PTC.TestInvariance(Y=SelData$d[,"ONECUT2", drop =F], X=ONECUT2.X.TScan)
ONECUT2.Parents<-Extract.Parents(ONECUT2.Parents,SelData$PParents["ONECUT2"])
ONECUT2.relMatrix<-InterList.toMatrix(ONECUT2.Parents$Names)
ONECUT2.Summary<-PTC.RankByContext(TS7.0_Conserved_Site_Context_Scores
                                  ,ONECUT2.relMatrix)
}
}
\seealso{
\link[PTC]{PTC}
}
